+++ src/layer/arm/convolution_sgemm_pack8to1_int8.h
@@ -1639,9 +1639,9 @@ static void convolution_im2col_sgemm_transform_kernel_pack8to1_int8_neon(const M
     // dst = 4a-4b-2-maxk-inch/8a-outch/4b (arm82)
     Mat kernel = _kernel.reshape(maxk, inch, outch);
     if (outch >= 4)
-        kernel_tm.create(32 * maxk, inch / 8, outch / 4 + outch % 4, 1u);
+        kernel_tm.create(32 * maxk, inch / 8, outch / 4 + outch % 4, (size_t)1u);
     else
-        kernel_tm.create(8 * maxk, inch / 8, outch, 1u);
+        kernel_tm.create(8 * maxk, inch / 8, outch, (size_t)1u);
 
     int q = 0;
     for (; q + 3 < outch; q += 4)
